requirements.txt
openai
python-dotenv
fastapi
uvicorn
pytest
pip install -r requirements.txt
